Introduce HFLoad for loading Pytorch and SafeTensor weights
