+++ python/mlc_chat/compiler/parameter/__init__.py
@@ -2,5 +2,5 @@
 A subpackage of the compiler that represents mapping between external parameters, quantized
 parameters and parameters in MLC-defined models.
 """
-from .hf_torch_loader import HFTorchLoader
+from .hf_loader import HFLoader
 from .mapping import ExternMapping, QuantizeMapping
